Refactor pilot/canvas.py and rename tests/test_pilot_canvas.py
